+++ nucliadb_dataset/CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 1.0.0a2 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Added get_info and get_ontology_count
 
 
 1.0.0a1 (2022-07-19)